Updated Login Controller tests Working J2ME Multi Forms with Obfuscation ViewFactory implemented for Login and Catalog Controllers

+++ ports/j2me/src/com/ustadmobile/app/tests/TestSystemimplJ2ME.java
@@ -67,7 +67,7 @@ public class TestSystemimplJ2ME extends TestCase {
         HTTPUtils.downloadURLToFile(url, getSharedContentDir, "");
         
         String fileURI2 = FileUtils.joinPath(getSharedContentDir, "hello.txt");
-        int modTimeDifference = ustadMobileSystemImpl.modTimeDifference(
+        int modTimeDifference = (int) ustadMobileSystemImpl.modTimeDifference(
                 fileURI2, fileURI);
         
         if ( modTimeDifference != -1 ){
